added count aggregation type. added color to scene graph 2d text.

+++ ev/endrov/windowPlateAnalysis/CalcAggregation.java
@@ -75,7 +75,7 @@ public class CalcAggregation
 		{
 		public Double calc(Collection<Double> listA, Collection<Double> listB)
 			{
-			return listA.size();
+			return (double)listA.size();
 			}
 		};
